fix(test_runner/truffle/truffle_runner): copy project directory with symlinks preserved
Truffle Solidity projects normally require installing dependencies via `npm`. When copying the project directory, the `npm` dependency directory `node_modules` is also copied over. When the symlinks are not preserved and resolved to the pointed files then copy, the files in `node_modules` that were symlinks will then be normal files. This is a problem when such files try to `require` other files using relative paths. The relative paths will be broken since those normal files are residing in places that they aren't designed to be.